Replace zoom with rebuilding and settings
The zoom in/zoom out wasn't very useful and you can always just open the
resume in an external window anyways to get the zoom tools if you
wanted. Instead, they've now been replaced with a button to rebuild the
resume and a button to open the settings menu (settings not implemented
yet).